First working version created
The logic behind the game is complete. It's definitely a little messy,
but I'm pretty happy with it considering using OOP in Ruby (and in general)
is quite new for me. I may go back and add an option to vs the computer
which would just make a random choice.